+++ src/viur_cli/utils.py
@@ -1,0 +1,11 @@
+import click
+
+
+def echo_error(msg):
+    """colored cli feedback"""
+    click.echo(click.style(msg, fg="red"))
+
+
+def echo_info(msg):
+    """colored cli feedback"""
+    click.echo(click.style(msg, fg="cyan"))