+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file documents any relevant changes.
 
+## [0.5.5] - 2021-12-13
+
+- [BREAKING]: `viur-cli` is now `viur`
+- [BREAKING]: `viur-cli deploy` is now `viur deploy app`
+- [BREAKING]: `viur-cli index` is now `viur deploy index`
+- [BREAKING]: `viur-cli flare [appname] [release | debug | watch]` is
+  now `viur flare [release | debug | watch] [appname]`
+
 ## [0.5.1] - 2021-12-03
 
 - fix: flare apps buildscripts